Started writing the Element class.

+++ src/Utilities/Display.hpp
@@ -3,14 +3,14 @@
 #ifndef Display_HPP
 #define Display_HPP
 
-void display(double *x, unsigned n)
+void display(float *x, unsigned n)
 {
     unsigned i;
     for(i=0; i<n; i++)
         printf("%6.2f\n",x[i]);
 }
 
-void display(double *A, unsigned m, unsigned n)
+void display(float *A, unsigned m, unsigned n)
 {
     unsigned i,j;
     for(i=0; i<m; i++)
