+++ src/Utilities/LobattoWeights.hpp
@@ -9,18 +9,18 @@
 #ifndef LobattoWeights_HPP
 #define LobattoWeights_HPP
 
-void lobattoWeights(double *Weights, unsigned N)
+void lobattoWeights(float *Weights, unsigned N)
 {
-    double *Poly, *Nodes;
-    Poly    =   new double[N];
-    Nodes   =   new double[N];
+    float *Poly, *Nodes;
+    Poly    =   new float[N];
+    Nodes   =   new float[N];
 	legendrePolynomial(Poly,N-1);
     lobattoNodes(Nodes,N);
 
-    function<double(double)> Eval;
+    function<float(float)> Eval;
     Weights[0]  = 2.0/((N)*(N-1));
 
-    Eval = [&Poly,&N](double x){ return(polyEval(Poly,N-1,x));};
+    Eval = [&Poly,&N](float x){ return(polyEval(Poly,N-1,x));};
 
 	for(int i=1;i<N-1;i++)
 		Weights[i]    =   2/((N*(N-1))*(Eval(Nodes[i]))*(Eval(Nodes[i])));
