+++ src/Utilities/Ones.hpp
@@ -4,14 +4,14 @@
 #ifndef Ones_HPP
 #define Ones_HPP
 
-void ones(double *x, unsigned n)
+void ones(float *x, unsigned n)
 {
     unsigned i;
     for(i=0; i<n; i++)
         x[i]    =   1.0;
 }
 
-void ones(double *A, unsigned m, unsigned n)
+void ones(float *A, unsigned m, unsigned n)
 {
     unsigned i;
     for(i=0;i<m*n;i++)
