+++ src/Utilities/Plot.hpp
@@ -8,7 +8,7 @@ using namespace std;
 #define Plot_HPP
 
 
-void plot(double *X, double *Y, double N, string GraphTitle = "", string LegendTitle = "", string fileName = "")
+void plot(float *X, float *Y, float N, string GraphTitle = "", string LegendTitle = "", string fileName = "")
 {
     unsigned i;
     FILE* pfile;
@@ -40,7 +40,7 @@ void plot(double *X, double *Y, double N, string GraphTitle = "", string LegendT
 /**
   * A similar 1-D plot but with the y-axis ranges.
   */
-void plot(double *X, double *Y, double N,double Ymin, double Ymax, string GraphTitle = "", string LegendTitle = "", string fileName = "")
+void plot(float *X, float *Y, float N,float Ymin, float Ymax, string GraphTitle = "", string LegendTitle = "", string fileName = "")
 {
     unsigned i;
     FILE* pfile;
@@ -76,7 +76,7 @@ void plot(double *X, double *Y, double N,double Ymin, double Ymax, string GraphT
  * @param Y    The Y-Domain.
  * @param Z    The Z-Domain.
  */
-void plot(double *X, double *Y, double *Z, unsigned Nx, unsigned Ny,double X1, double X2,double Y1, double Y2,double Z1, double Z2, string Name)
+void plot(float *X, float *Y, float *Z, unsigned Nx, unsigned Ny,float X1, float X2,float Y1, float Y2,float Z1, float Z2, string Name)
 {
     int i;
     FILE* pfile;
