+++ src/Utilities/PolyDeriv.hpp
@@ -3,7 +3,7 @@
 #ifndef PolyDeriv_HPP
 #define PolyDeriv_HPP
 
-void polyDeriv(double *GivenPoly, double *DerivedPoly, unsigned deg)
+void polyDeriv(float *GivenPoly, float *DerivedPoly, unsigned deg)
 {
 	for(int i = 1;i<=deg;i++)
         DerivedPoly[i-1] = i*GivenPoly[i];
