+++ src/Utilities/Zeros.hpp
@@ -3,14 +3,14 @@
 #ifndef Zeros_HPP
 #define Zeros_HPP
 
-void zeros(double *x, unsigned n)
+void zeros(float *x, unsigned n)
 {
     unsigned i;
     for(i=0; i<n; i++)
         x[i]    =   0.0;
 }
 
-void zeros(double *A, unsigned m, unsigned n)
+void zeros(float *A, unsigned m, unsigned n)
 {
     unsigned i;
     for(i=0;i<m*n;i++)
